update build script

+++ build_kernel.sh
@@ -6,8 +6,14 @@ export CROSS_COMPILE=~/arm32/bin/arm-eabi-
 mkdir output
 
 #make -C $(pwd) O=output msm8974_sec_defconfig VARIANT_DEFCONFIG=msm8974pro_sec_klte_eur_defconfig SELINUX_DEFCONFIG=selinux_defconfig
-cp defconfig output/.config
-make -j4 O=output oldconfig
+if [ -e output/.config ]
+    then
+        make -j4 O=output oldconfig
+    else
+        cp defconfig output/.config
+        make -j4 O=output oldconfig
+fi
+
 make -j4 O=output
 
 cp output/arch/arm/boot/Image $(pwd)/arch/arm/boot/zImage
